feat: add functionality for the files in the modules and zip filders
