+++ src/zip/compress.js
@@ -1,3 +1,15 @@
+import fs from 'fs'
+import zlib from 'zlib'
+
 export const compress = async () => {
-    // Write your code here 
+    
+let readableStream = fs.createReadStream("./files/fileToCompress.txt", "utf8");
+  
+let writeableStream = fs.createWriteStream("./files/archive.gz");
+  
+let gzip = zlib.createGzip();
+  
+readableStream.pipe(gzip).pipe(writeableStream);
 };
+
+compress();
